+++ x/fswap/genesis_test.go
@@ -1,1 +1,0 @@
-package fswap_test