+++ x/fswap/types/keys.go
@@ -17,6 +17,11 @@ const (
 	MemStoreKey = "mem_fswap"
 )
 
+var (
+	ParamsKey  = byte(0x00)
+	SwappedKey = byte(0x01)
+)
+
 func KeyPrefix(p string) []byte {
 	return []byte(p)
 }